refactor: 重新设计 Step 组件
