+++ src/BootstrapBlazor/Components/Step/Step.razor.cs
@@ -5,118 +5,40 @@
 namespace BootstrapBlazor.Components;
 
 /// <summary>
-/// Step 组件
+/// Step 组件类
 /// </summary>
-public sealed partial class Step
+public partial class Step
 {
-    private string? ClassString => CssBuilder.Default("step is-horizontal")
-        .AddClass("is-flex", IsLast && !((Steps?.IsCenter ?? false) || IsCenter))
-        .AddClass("is-center", (Steps?.IsCenter ?? false) || IsCenter)
+    /// <summary>
+    /// 获得 组件样式字符串
+    /// </summary>
+    private string? ClassString => CssBuilder.Default("step")
+        .AddClassFromAttributes(AdditionalAttributes)
         .Build();
 
-    private string? StyleString => CssBuilder.Default("margin-right: 0px;")
-        .AddClass($"flex-basis: {Space};", !string.IsNullOrEmpty(Space))
+    private string? HeaderClassString => CssBuilder.Default("step-header")
+        .AddClass("steps-horizontal", !IsVertical)
+        .AddClass("steps-vertical", IsVertical)
         .Build();
 
-    private string? HeadClassString => CssBuilder.Default("step-head")
-        .AddClass($"is-{Status.ToDescriptionString()}")
-        .Build();
+    //private static string? GetContentClassString(StepOption item) => CssBuilder.Default("step-body")
+    //    .AddClass("d-none", !item.IsActive)
+    //    .Build();
 
-    private string? LineStyleString => CssBuilder.Default()
-        .AddClass("transition-delay: 150ms; border-width: 1px; width: 100%;", Status == StepStatus.Finish || Status == StepStatus.Success)
-        .Build();
-
-    private string? StepIconClassString => CssBuilder.Default("step-icon")
-        .AddClass("is-text", !IsIcon)
-        .AddClass("is-icon", IsIcon)
-        .Build();
-
-    private string? IconClassString => CssBuilder.Default("step-icon-inner")
-        .AddClass(Icon, IsIcon || Status == StepStatus.Finish || Status == StepStatus.Success)
-        .AddClass(ErrorStepIcon, IsIcon || Status == StepStatus.Error)
-        .AddClass("is-status", !IsIcon && (Status == StepStatus.Finish || Status == StepStatus.Success || Status == StepStatus.Error))
-        .Build();
-
-    private string? TitleClassString => CssBuilder.Default("step-title")
-        .AddClass($"is-{Status.ToDescriptionString()}")
-        .Build();
-
-    private string? DescClassString => CssBuilder.Default("step-description")
-        .AddClass($"is-{Status.ToDescriptionString()}")
-        .Build();
-
-    private string? StepString => (Status == StepStatus.Process || Status == StepStatus.Wait) && !IsIcon ? (StepIndex + 1).ToString() : null;
+    //private StepOption? CurrentStep { get; set; }
 
     /// <summary>
-    /// 获得/设置 步骤显示文字
+    /// 获得/设置 步骤集合
     /// </summary>
     [Parameter]
-    public string? Title { get; set; }
+    [NotNull]
+    public List<StepOption>? Items { get; set; }
 
     /// <summary>
-    /// 获得/设置 步骤显示图标
+    /// 获得/设置 是否垂直渲染 默认 false 水平渲染
     /// </summary>
     [Parameter]
-    public string? Icon { get; set; }
-
-    /// <summary>
-    /// 获得/设置 步骤显示图标
-    /// </summary>
-    public string? ErrorIcon { get; set; }
-
-    /// <summary>
-    /// 获得/设置 错误步骤显示图标
-    /// </summary>
-    [Parameter]
-    public string? ErrorStepIcon { get; set; }
-
-    /// <summary>
-    /// 获得/设置 步骤状态
-    /// </summary>
-    [Parameter]
-    public StepStatus Status { get; set; }
-
-    /// <summary>
-    /// 获得/设置 描述信息
-    /// </summary>
-    [Parameter]
-    public string? Description { get; set; }
-
-    /// <summary>
-    /// 获得/设置 step 的间距不填写将自适应间距支持百分比
-    /// </summary>
-    [Parameter]
-    public string? Space { get; set; }
-
-    /// <summary>
-    /// 获得/设置 是否为图标
-    /// </summary>
-    [Parameter]
-    public bool IsIcon { get; set; }
-
-    /// <summary>
-    /// 获得/设置 是否为最后一个 Step
-    /// </summary>
-    [Parameter]
-    public bool IsLast { get; set; }
-
-    /// <summary>
-    /// 获得/设置 是否居中对齐
-    /// </summary>
-    [Parameter]
-    public bool IsCenter { get; set; }
-
-    /// <summary>
-    /// 获得/设置 Step 顺序
-    /// </summary>
-    [Parameter]
-    public int StepIndex { get; set; }
-
-    /// <summary>
-    /// 获得/设置 父级组件 Steps 实例
-    /// </summary>
-    [CascadingParameter]
-    private Steps? Steps { get; set; }
+    public bool IsVertical { get; set; }
 
     /// <summary>
     /// 获得/设置 步骤组件状态改变时回调委托
@@ -130,9 +52,57 @@ public sealed partial class Step
     [Parameter]
     public RenderFragment? DescriptionTemplate { get; set; }
 
-    [Inject]
-    [NotNull]
-    private IIconTheme? IconTheme { get; set; }
+    /// <summary>
+    /// 获得/设置 组件内容实例
+    /// </summary>
+    private RenderFragment? ChildContent { get; set; }
+
+    ///// <summary>
+    ///// 获得/设置 步骤组件状态改变时回调委托
+    ///// </summary>
+    //[Parameter]
+    //public Func<StepStatus, Task>? OnStatusChanged { get; set; }
+
+    //[Inject]
+    //[NotNull]
+    //private IIconTheme? IconTheme { get; set; }
+
+    //private string? HeadClassString => CssBuilder.Default("step-head")
+    //    .AddClass($"is-{Status.ToDescriptionString()}")
+    //    .Build();
+
+    //private string? LineStyleString => CssBuilder.Default()
+    //    .AddClass("width: 100%;", Status == StepStatus.Finish || Status == StepStatus.Success)
+    //    .Build();
+
+    //private static string? GetStepIconClassString(StepOption option) => CssBuilder.Default("step-icon")
+    //    .AddClass("is-text", !option.IsIcon)
+    //    .AddClass("is-icon", option.IsIcon)
+    //    .Build();
+
+    //private string? IconClassString => CssBuilder.Default("step-icon-inner")
+    //    .AddClass(Icon, IsIcon || Status == StepStatus.Finish || Status == StepStatus.Success)
+    //    .AddClass(ErrorStepIcon, IsIcon || Status == StepStatus.Error)
+    //    .AddClass("is-status", !IsIcon && (Status == StepStatus.Finish || Status == StepStatus.Success || Status == StepStatus.Error))
+    //    .Build();
+
+    //private string? TitleClassString => CssBuilder.Default("step-title")
+    //    .AddClass($"is-{Status.ToDescriptionString()}")
+    //    .Build();
+
+    //private string? DescClassString => CssBuilder.Default("step-description")
+    //    .AddClass($"is-{Status.ToDescriptionString()}")
+    //    .Build();
+
+    private string GetStepString(StepOption option) => $"{Items.IndexOf(option) + 1}";
+
+    private static string? GetStepItemClassString(StepOption option) => CssBuilder.Default("step-body-item")
+        .AddClass("d-none", !option.IsActive)
+        .Build();
+
+    private static string? GetIconClassString(StepOption option) => CssBuilder.Default("step-icon")
+        .AddClass(option.Icon)
+        .Build();
 
     /// <summary>
     /// <inheritdoc/>
@@ -141,7 +111,151 @@ public sealed partial class Step
     {
         base.OnParametersSet();
 
-        Icon ??= IconTheme.GetIconByKey(ComponentIcons.StepIcon);
-        ErrorIcon = IconTheme.GetIconByKey(ComponentIcons.StepErrorIcon);
+        Items ??= new();
+    }
+
+    ///// <summary>
+    ///// OnParametersSetAsync 方法
+    ///// </summary>
+    //protected override async Task OnParametersSetAsync()
+    //{
+    //    await base.OnParametersSetAsync();
+
+    //    CurrentStep = null;
+    //    if (Items.Any())
+    //    {
+    //        CurrentStep = Items.FirstOrDefault(i => i.IsActive);
+    //        var status = CurrentStep?.Status ?? StepStatus.Wait;
+    //        if (Status != status)
+    //        {
+    //            Status = status;
+    //            if (OnStatusChanged != null)
+    //            {
+    //                await OnStatusChanged.Invoke(Status);
+    //            }
+    //        }
+    //    }
+    //}
+
+    ///// <summary>
+    ///// 渲染 Step 组件方法
+    ///// </summary>
+    ///// <param name="item"></param>
+    ///// <returns></returns>
+    //protected virtual RenderFragment RenderStep(StepOption item) => new(builder =>
+    //{
+    //    item.Space = ParseSpace(item.Space);
+    //    var index = 0;
+    //    builder.OpenComponent<Step>(index++);
+    //    builder.SetKey(item);
+    //    builder.AddAttribute(index++, nameof(Step.Title), item.Title);
+    //    builder.AddAttribute(index++, nameof(Step.Icon), item.Icon);
+    //    builder.AddAttribute(index++, nameof(Step.Description), item.Description);
+    //    builder.AddAttribute(index++, nameof(Step.Space), item.Space);
+    //    builder.AddAttribute(index++, nameof(Step.Status), item.Status);
+    //    builder.AddAttribute(index++, nameof(Step.IsLast), item == Items.Last());
+    //    builder.AddAttribute(index++, nameof(Step.IsCenter), IsCenter);
+    //    builder.AddAttribute(index, nameof(Step.StepIndex), Items.ToList().IndexOf(item));
+    //    builder.CloseComponent();
+    //});
+
+    //private string ParseSpace(string? space)
+    //{
+    //    if (!string.IsNullOrEmpty(space) && !double.TryParse(space.TrimEnd('%'), out _)) space = null;
+    //    if (string.IsNullOrEmpty(space)) space = $"{Math.Round(100 * 1.0d / Math.Max(1, Items.Count - 1), 2)}%";
+    //    return space;
+    //}
+
+    /// <summary>
+    /// 上一步
+    /// </summary>
+    /// <param name="status">当前要设定的状态</param>
+    public void Prev(StepStatus status = StepStatus.Wait)
+    {
+        var item = Items.Find(x => x.IsActive);
+        if (item != null)
+        {
+            var index = Items.IndexOf(item);
+            if (index <= 0)
+            {
+                return;
+            }
+
+            index--;
+            item.IsActive = false;
+            item.Status = status;
+            item = Items[index];
+            item.Status = StepStatus.Process;
+            item.IsActive = true;
+            StateHasChanged();
+            OnStatusChanged?.Invoke(status);
+        }
+        else
+        {
+            item = Items.LastOrDefault();
+            if (item == null)
+            {
+                return;
+            }
+            item.IsActive = true;
+            item.Status = StepStatus.Process;
+            StateHasChanged();
+        }
+    }
+
+    /// <summary>
+    /// 下一步
+    /// </summary>
+    /// <param name="status">当前要设定的状态</param>
+    public void Next(StepStatus status = StepStatus.Success)
+    {
+        var item = Items.Find(x => x.IsActive);
+        if (item != null)
+        {
+            var index = Items.IndexOf(item);
+            if (index >= Items.Count)
+            {
+                return;
+            }
+            item.IsActive = false;
+            item.Status = status;
+            index++;
+            if (index < Items.Count)
+            {
+                item = Items[index];
+                item.Status = StepStatus.Process;
+                item.IsActive = true;
+            }
+            StateHasChanged();
+            OnStatusChanged?.Invoke(status);
+        }
+    }
+
+    /// <summary>
+    /// 添加步骤到组件中
+    /// </summary>
+    /// <param name="option"></param>
+    public void Add(StepOption option)
+    {
+        Items.Add(option);
+    }
+
+    /// <summary>
+    /// 插入步骤到组件中
+    /// </summary>
+    /// <param name="index"></param>
+    /// <param name="option"></param>
+    public void Insert(int index, StepOption option)
+    {
+        Items.Insert(index, option);
+    }
+
+    /// <summary>
+    /// 从组件中移除步骤
+    /// </summary>
+    /// <param name="option"></param>
+    public void Remove(StepOption option)
+    {
+        Items.Remove(option);
     }
 }
