--- src/BootstrapBlazor/Components/Step/Steps.razor
@@ -1,9 +1,0 @@
-﻿@namespace BootstrapBlazor.Components
-@inherits BootstrapComponentBase
-@attribute [ExcludeFromCodeCoverage]
-
-<div @attributes="@AdditionalAttributes" class="@ClassString">
-    <CascadingValue Value="this" IsFixed="true">
-        @ChildContent
-    </CascadingValue>
-</div>
